+++ lib/src/config_provider/config.dart
@@ -181,12 +181,25 @@ class Config {
   /// Key: Name, Value: [Specification]
   Map<List<String>, Specification> _getSpecs() {
     return <List<String>, Specification>{
+      //TODO: Deprecated, remove in next major update.
       [strings.llvmLib]: Specification<String>(
         requirement: Requirement.no,
         validator: llvmLibValidator,
         extractor: llvmLibExtractor,
+        defaultValue: () => '',
+        extractedResult: (dynamic result) {
+          _libclangDylib = result as String;
+        },
+      ),
+      [strings.llvmPath]: Specification<String>(
+        requirement: Requirement.no,
+        validator: llvmPathValidator,
+        extractor: llvmPathExtractor,
         defaultValue: () => findDylibAtDefaultLocations(),
-        extractedResult: (dynamic result) => _libclangDylib = result as String,
+        extractedResult: (dynamic result) {
+          // If this key wasn't already set by `llvm-lib` use this result.
+          if (_libclangDylib.isEmpty) _libclangDylib = result as String;
+        },
       ),
       [strings.output]: Specification<String>(
         requirement: Requirement.yes,