+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 2.4.0
+- Added new config key `llvm-path` that accepts a list of `path/to/llvm`.
+- Deprecated config key `llvm-lib`.
+
 # 2.3.0
 - Added config key `compiler-opts-automatic -> macos -> include-c-standard-library`
 (default: true) to automatically find and add C standard library on macOS.